backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# AI/ML - Speech to Text
openai-whisper>=20231117

# AI/ML - Embeddings & Vision
transformers>=4.35.0
sentence-transformers>=2.2.2
torch>=2.1.0
torchvision>=0.16.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Vector Database
qdrant-client>=1.6.0

# Video Processing
yt-dlp>=2023.11.0

# Background Tasks
celery>=5.3.0
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.26.4
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
ipykernel>=6.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
